Manual low voltage power off
Trying to keep the RTC alive. Seems to be loosing time when the battery
dies.

Wasn't doing this before but the voltage coming into this pin was too
high and caused a chip to blow (even though it kept time beautifully
until the end). Looks like the safer lower voltage can get too low
however.